+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/masked_bincount/CMakeLists.txt
@@ -1,0 +1,54 @@
+add_library(ttnn_op_experimental_deepseek_prefill_masked_bincount ${LIB_TYPE})
+add_library(
+    TTNN::Ops::Experimental::DeepSeekPrefill::MaskedBincount
+    ALIAS ttnn_op_experimental_deepseek_prefill_masked_bincount
+)
+
+target_precompile_headers(ttnn_op_experimental_deepseek_prefill_masked_bincount REUSE_FROM TTNN::PCH)
+TT_ENABLE_UNITY_BUILD(ttnn_op_experimental_deepseek_prefill_masked_bincount)
+
+set_target_properties(
+    ttnn_op_experimental_deepseek_prefill_masked_bincount
+    PROPERTIES
+        INTERFACE_HEADER_SETS_TO_VERIFY
+            api
+)
+
+file(GLOB_RECURSE kernels device/kernels/*)
+
+target_sources(
+    ttnn_op_experimental_deepseek_prefill_masked_bincount
+    PUBLIC
+        FILE_SET api
+        TYPE HEADERS
+        BASE_DIRS ${FixmeOpAPIDir}
+        FILES
+        FILE_SET kernels
+        TYPE HEADERS
+        BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}
+        FILES ${kernels}
+    PRIVATE
+        device/masked_bincount_device_operation.cpp
+        device/masked_bincount_program_factory.cpp
+        masked_bincount.cpp
+)
+
+target_include_directories(ttnn_op_experimental_deepseek_prefill_masked_bincount PRIVATE ${FixmeOpIncDirs})
+target_link_libraries(
+    ttnn_op_experimental_deepseek_prefill_masked_bincount
+    PRIVATE
+        TT::Metalium
+        TTNN::Core
+)
+
+install(
+    TARGETS
+        ttnn_op_experimental_deepseek_prefill_masked_bincount
+    FILE_SET
+    kernels
+        DESTINATION
+            ${CMAKE_INSTALL_LIBEXECDIR}/tt-metalium/ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/masked_bincount
+        COMPONENT ttnn-runtime
+)
+
+install(TARGETS ttnn_op_experimental_deepseek_prefill_masked_bincount LIBRARY COMPONENT tar)
