+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/offset_cumsum/CMakeLists.txt
@@ -1,0 +1,54 @@
+add_library(ttnn_op_experimental_deepseek_prefill_offset_cumsum ${LIB_TYPE})
+add_library(
+    TTNN::Ops::Experimental::DeepSeekPrefill::OffsetCumsum
+    ALIAS ttnn_op_experimental_deepseek_prefill_offset_cumsum
+)
+
+target_precompile_headers(ttnn_op_experimental_deepseek_prefill_offset_cumsum REUSE_FROM TTNN::PCH)
+TT_ENABLE_UNITY_BUILD(ttnn_op_experimental_deepseek_prefill_offset_cumsum)
+
+set_target_properties(
+    ttnn_op_experimental_deepseek_prefill_offset_cumsum
+    PROPERTIES
+        INTERFACE_HEADER_SETS_TO_VERIFY
+            api
+)
+
+file(GLOB_RECURSE kernels device/kernels/*)
+
+target_sources(
+    ttnn_op_experimental_deepseek_prefill_offset_cumsum
+    PUBLIC
+        FILE_SET api
+        TYPE HEADERS
+        BASE_DIRS ${FixmeOpAPIDir}
+        FILES
+        FILE_SET kernels
+        TYPE HEADERS
+        BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}
+        FILES ${kernels}
+    PRIVATE
+        device/offset_cumsum_device_operation.cpp
+        device/offset_cumsum_program_factory.cpp
+        offset_cumsum.cpp
+)
+
+target_include_directories(ttnn_op_experimental_deepseek_prefill_offset_cumsum PRIVATE ${FixmeOpIncDirs})
+target_link_libraries(
+    ttnn_op_experimental_deepseek_prefill_offset_cumsum
+    PRIVATE
+        TT::Metalium
+        TTNN::Core
+)
+
+install(
+    TARGETS
+        ttnn_op_experimental_deepseek_prefill_offset_cumsum
+    FILE_SET
+    kernels
+        DESTINATION
+            ${CMAKE_INSTALL_LIBEXECDIR}/tt-metalium/ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/offset_cumsum
+        COMPONENT ttnn-runtime
+)
+
+install(TARGETS ttnn_op_experimental_deepseek_prefill_offset_cumsum LIBRARY COMPONENT tar)
