+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/offset_cumsum/offset_cumsum_nanobind.hpp
@@ -6,10 +6,10 @@
 
 #include "ttnn-nanobind/nanobind_fwd.hpp"
 
-namespace ttnn::operations::experimental::offset_cumsum::detail {
+namespace ttnn::operations::experimental::deepseek_prefill::offset_cumsum::detail {
 
 namespace nb = nanobind;
 
 void bind_experimental_offset_cumsum_operation(nb::module_& mod);
 
-}  // namespace ttnn::operations::experimental::offset_cumsum::detail
+}  // namespace ttnn::operations::experimental::deepseek_prefill::offset_cumsum::detail
