+++ ttnn/cpp/ttnn/operations/experimental/experimental_nanobind.cpp
@@ -62,8 +62,8 @@
 #include "ttnn/operations/experimental/deepseek_prefill/dispatch/dispatch_nanobind.hpp"
 #include "ttnn/operations/experimental/deepseek_prefill/combine/combine_nanobind.hpp"
 #include "ttnn/operations/experimental/deepseek_moe_post_combine_tilize/deepseek_moe_post_combine_tilize_nanobind.hpp"
-#include "ttnn/operations/experimental/masked_bincount/masked_bincount_nanobind.hpp"
-#include "ttnn/operations/experimental/offset_cumsum/offset_cumsum_nanobind.hpp"
+#include "ttnn/operations/experimental/deepseek_prefill/masked_bincount/masked_bincount_nanobind.hpp"
+#include "ttnn/operations/experimental/deepseek_prefill/offset_cumsum/offset_cumsum_nanobind.hpp"
 
 namespace ttnn::operations::experimental {
 
@@ -120,8 +120,8 @@ void py_module(nb::module_& mod) {
     matmul::detail::bind_attn_matmul(mod);
     matmul::detail::bind_attn_matmul_from_cache(mod);
     matmul::detail::bind_group_attn_matmul(mod);
-    masked_bincount::detail::bind_experimental_masked_bincount_operation(mod);
-    offset_cumsum::detail::bind_experimental_offset_cumsum_operation(mod);
+    deepseek_prefill::masked_bincount::detail::bind_experimental_masked_bincount_operation(mod);
+    deepseek_prefill::offset_cumsum::detail::bind_experimental_offset_cumsum_operation(mod);
 
     plusone::detail::bind_experimental_plusone_operation(mod);
     dropout::detail::bind_experimental_dropout_operation(mod);
